use ImageFormat to load and save images correctly. Saving animated images in their origin file format, non-animated as pngs if images have transparency, jpeg if not - as before. Full support of apng files now
