+++ deltachat-ios/Helper/ImageFormat.swift
@@ -1,6 +1,7 @@
 import Foundation
 import UIKit
 import SDWebImage
+import DcCore
 
 enum ImageFormat: String {
     case png, jpg, gif, tiff, webp, heic, bmp, unknown
@@ -67,4 +68,35 @@ extension ImageFormat {
         return loadImageFrom(data: imageData)
     }
 
+    public static func saveImage(image: UIImage) -> String? {
+        if image.sd_isAnimated,
+           let data = image.sd_imageData() {
+            let format = ImageFormat.get(from: data)
+            if format != .unknown {
+                return ImageFormat.saveImage(data: data, suffix: format.rawValue)
+            }
+        }
+        let suffix = image.isTransparent() ? "png" : "jpg"
+        guard let data = image.isTransparent() ? image.pngData() : image.jpegData(compressionQuality: 1.0) else {
+            return nil
+        }
+
+        return saveImage(data: data, suffix: suffix)
+    }
+
+    public static func saveImage(data: Data, suffix: String) -> String? {
+        let timestamp = Double(Date().timeIntervalSince1970)
+        guard let directory = try? FileManager.default.url(for: .documentDirectory, in: .userDomainMask,
+                                                           appropriateFor: nil, create: false) as NSURL,
+            let path = directory.appendingPathComponent("\(timestamp).\(suffix)")
+            else { return nil }
+
+        do {
+            try data.write(to: path)
+            return path.relativePath
+        } catch {
+            DcContext.shared.logger?.info(error.localizedDescription)
+            return nil
+        }
+    }
 }
